fix: relocate chart flies
