+++ src/components/charts/templates/RainFallAreaChart.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import {
   ResponsiveContainer,
-  ComposedChart,
+  AreaChart,
   CartesianGrid,
   XAxis,
   YAxis,
@@ -10,22 +10,16 @@ import {
   Area,
   Line
 } from "recharts";
-import ChartCard from '../../../components/ChartCard/ChartCard';
+import ChartCard from '../chartCustoms/ChartCard';
 import { RAINFALL } from '../../../data/melb-monthly-rainfail';
-import { TEMPERATURE } from '../../../data/melb-monthly-temperature';
 import { getMonthNameByOrder } from '../../../monthMapping';
 import { TooltipContainerStyles } from "../../../styles/constants/tooltipContainerStyles";
 
-const RainFallTemperatureComposedChart: React.FC = () => {
-  const data = RAINFALL["2019"].map((rainfall, index) => ({
-    ...rainfall,
-    temperature: TEMPERATURE["2019"][index].temperature
-  }));
-
+const RainFallAreaChart: React.FC = () => {
   return (
-    <ChartCard heading="Melbourne 2019 monthly rainfall and temperature">
-      <ResponsiveContainer width="100%" height={220}>
-        <ComposedChart data={data} style={{ fontWeight: 'bold' }}>
+    <ChartCard heading="Melbourne 2019 monthly rainfall">
+      <ResponsiveContainer width="100%" height={200}>
+        <AreaChart data={RAINFALL["2019"]} style={{ fontWeight: 'bold' }}>
           <defs>
             <linearGradient id="rainGradient" x1="0" y1="0" x2="0" y2="1">
               <stop offset="0%" stopColor="#3066BE" />
@@ -37,23 +31,10 @@ const RainFallTemperatureComposedChart: React.FC = () => {
             strokeDasharray="3 3"
             stroke="#d6d9da"
           />
-          <XAxis
-            dataKey="month"
-            scale="point"
-          />
+          <XAxis dataKey="month" />
           <YAxis
-            yAxisId="rainfall"
             unit="ml"
             orientation="left"
-            width={35}
-            axisLine={false}
-            tickLine={false}
-          />
-          <YAxis
-            yAxisId="temperature"
-            unit="°C"
-            orientation="right"
-            domain={["auto", "auto"]}
             width={35}
             axisLine={false}
             tickLine={false}
@@ -62,30 +43,18 @@ const RainFallTemperatureComposedChart: React.FC = () => {
             labelFormatter={getMonthNameByOrder}
             cursor={false}
             contentStyle={TooltipContainerStyles}
-            itemStyle={{ paddingBottom: 0 }}
           />
           <Area
-            yAxisId="rainfall"
             dataKey="rainfall"
             name="Rainfall"
             unit="ml"
             type="basis"
             fill="url(#rainGradient)"
           />
-          <Line
-            yAxisId="temperature"
-            dataKey="temperature"
-            name="Temperature"
-            type="monotone"
-            stroke="#EF5B5B"
-            fill="#EF5B5B"
-            unit="°C"
-          />
-          <Legend />
-        </ComposedChart>
+        </AreaChart>
       </ResponsiveContainer>
     </ChartCard>
   );
 };
 
-export default RainFallTemperatureComposedChart;
+export default RainFallAreaChart;
