+++ src/routes/routes.tsx
@@ -1,4 +1,3 @@
-import Charts from "./Charts/Charts";
 import { BrowserRouter } from "react-router-dom";
 import { Route, Routes } from "react-router-dom";
 import Dashboard from "../pages/Dashboard";
@@ -7,10 +6,12 @@ import CreateCampagin from "../pages/CreateCampagin";
 import EditCampagin from "../pages/EditCampagin";
 import { ThemeProvider } from "@mui/material";
 import { theme } from "../styles/theme";
+import { RecoilRoot } from 'recoil';
 
 const Router = () => {
   return (
     <BrowserRouter>
+    <RecoilRoot>
       <ThemeProvider theme={theme}>
         <Routes>
           <Route path="/" element={<Dashboard />} />
@@ -18,8 +19,8 @@ const Router = () => {
           <Route path="/campagin-create" element={<CreateCampagin />} />
           <Route path="/campagin-edit" element={<EditCampagin />} />
         </Routes>
-        {/* <Charts /> */}
       </ThemeProvider>
+      </RecoilRoot>
     </BrowserRouter>
   );
 };
