+++ src/store/charts.ts
@@ -1,0 +1,13 @@
+import { atom, selectorFamily } from 'recoil';
+
+export const chartFirstData = atom({
+  key: 'chartFirstData',
+  default: "",
+});
+
+export const chartSecondData = atom({
+  key: 'chartSecondData',
+  default: "",
+});
+
+
